requirements.txt
Django==4.2.14
djangorestframework==3.15.2
django-environ==0.11.2
django-crontab==0.7.1
requests==2.32.3
psycopg2-binary==2.9.9
whitenoise==6.6.0